use doubles instead of floats

+++ enemy.h
@@ -11,13 +11,13 @@
 typedef struct Enemy {
     SDL_Texture *texture;
     SDL_FRect    rect;
-    float        speed;
-    float        rotation;       // hitbox/draw rotation in degrees
-    Vec2f        move_direction; // unit vector
+    double       speed;
+    double       rotation;       // hitbox/draw rotation in degrees
+    Vec2d        move_direction; // unit vector
 } Enemy;
 
-void Enemy_get_points(const Enemy *this, Vec2f dst[4]);
-void Enemy_get_normals(const Enemy *this, Vec2f dst[2]);
+void Enemy_get_points(const Enemy *this, Vec2d dst[4]);
+void Enemy_get_normals(const Enemy *this, Vec2d dst[2]);
 
 int  spawn_enemy(Enemy *dst);
 bool collide(Player *player, Enemy *enemy);
